+++ examples/command.rs
@@ -14,14 +14,14 @@ async fn main() -> OpenRgbResult<()> {
     println!("Performing example on {}", controller.name());
     let mut cmd = controller.cmd();
     // Set all LEDs to red
-    cmd.add_set_leds(vec![Color::new(255, 0, 0); controller.num_leds()])?;
+    cmd.set_leds(vec![Color::new(255, 0, 0); controller.num_leds()])?;
     // First zone to green
-    cmd.add_set_zone_leds(
+    cmd.set_zone_leds(
         0,
         vec![Color::new(0, 255, 0); controller.get_zone(0)?.num_leds()],
     )?;
     // First led to blue
-    cmd.add_set_led(0, Color::new(0, 0, 255))?;
+    cmd.set_led(0, Color::new(0, 0, 255))?;
     // This is now equivalent to a single `controller.set_leds(...)` command
     cmd.execute().await?;