+++ examples/command_group.rs
@@ -7,9 +7,12 @@ async fn main() -> OpenRgbResult<()> {
     let group = client.get_all_controllers().await?;
     let mut cmd = group.cmd();
     for c in &group {
-        cmd.add_update_controller_leds(c, vec![Color::new(255, 0, 0); c.num_leds()])?;
+        cmd.set_controller_leds(c, vec![Color::new(255, 0, 0); c.num_leds()])?;
     }
     cmd.execute().await?;
+
+    let first = group.get_controller(0)?;
+    first.set_all_leds(Color::new(255, 255, 255)).await?;
 
     Ok(())
 }