+++ src/client/command.rs
@@ -7,19 +7,15 @@ use crate::{
 
 /// The different type of LED updates that are possible
 ///
-/// For use with `UpdateLedCommand` and `UpdateLedCommandGroup`.
-pub enum UpdateCommand {
+/// For use with [`Command`] and [`CommandGroup`].
+enum SetLedCommand {
     /// Update LEDs in a controller
     Controller {
-        /// The ID of the controller
-        controller_id: usize,
         /// The colors to set
         colors: Vec<Color>,
     },
     /// Update LEDs in a zone in a controller
     Zone {
-        /// The ID of the controller
-        controller_id: usize,
         /// The ID of the zone
         zone_id: usize,
         /// The colors to set in the zone
@@ -27,8 +23,6 @@ pub enum UpdateCommand {
     },
     /// Update LEDs in a segment in a zone in a controller
     Segment {
-        /// The ID of the controller
-        controller_id: usize,
         /// The ID of the zone
         zone_id: usize,
         /// The ID of the segment
@@ -38,8 +32,6 @@ pub enum UpdateCommand {
     },
     /// Update a single LED in a controller
     Single {
-        /// The ID of the controller
-        controller_id: usize,
         /// LED index in the controller
         led_id: usize,
         /// The color to set the LED to
@@ -47,20 +39,20 @@ pub enum UpdateCommand {
     },
 }
 
-/// An `UpdateLedCommand` for a `ControllerGroup`, which allow you to update multiple controllers.
+/// [`Command`] equivalent for a [`ControllerGroup`], which allow you to update multiple controllers.
 ///
 /// This is useful when doing updates for multiple controllers at once.
-pub struct UpdateLedCommandGroup<'a> {
+pub struct CommandGroup<'a> {
     group: &'a ControllerGroup,
-    commands: HashMap<usize, UpdateLedCommand<'a>>,
+    commands: HashMap<usize, Command<'a>>,
 }
 
-impl<'a> UpdateLedCommandGroup<'a> {
+impl<'a> CommandGroup<'a> {
     pub(crate) fn new(group: &'a ControllerGroup) -> Self {
         let map = group
             .controllers()
             .iter()
-            .map(|c| (c.id(), UpdateLedCommand::new(c)))
+            .map(|c| (c.id(), Command::new(c)))
             .collect();
         Self {
             group,
@@ -76,10 +68,11 @@ impl<'a> UpdateLedCommandGroup<'a> {
         Ok(())
     }
 
-    fn get_controller_mut(
+    /// Returns a mutable reference to the command for the given controller ID.
+    fn get_cmd_mut(
         &mut self,
         controller_id: impl ControllerIndex,
-    ) -> OpenRgbResult<&mut UpdateLedCommand<'a>> {
+    ) -> OpenRgbResult<&mut Command<'a>> {
         let c = self.group.get_controller(controller_id)?;
         self.commands
             .get_mut(&c.id())
@@ -94,14 +87,14 @@ impl<'a> UpdateLedCommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn add_update_led(
+    pub fn set_controller_led(
         &mut self,
         controller_id: impl ControllerIndex,
         led_id: usize,
         color: Color,
     ) -> OpenRgbResult<()> {
-        let cmd = self.get_controller_mut(controller_id)?;
-        cmd.add_set_led(led_id, color)
+        let cmd = self.get_cmd_mut(controller_id)?;
+        cmd.set_led(led_id, color)
     }
 
     /// Add a command to update a multiple LEDs in a controller
@@ -109,13 +102,13 @@ impl<'a> UpdateLedCommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn add_update_controller_leds(
+    pub fn set_controller_leds(
         &mut self,
         controller_id: impl ControllerIndex,
-        colors: Vec<Color>,
+        colors: impl IntoIterator<Item = Color>,
     ) -> OpenRgbResult<()> {
-        let cmd = self.get_controller_mut(controller_id)?;
-        cmd.add_set_leds(colors)
+        let cmd = self.get_cmd_mut(controller_id)?;
+        cmd.set_leds(colors)
     }
 
     /// Add a command to update a zone in a controller.
@@ -123,14 +116,14 @@ impl<'a> UpdateLedCommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn add_update_zone(
+    pub fn set_controller_zone_leds(
         &mut self,
         controller_id: impl ControllerIndex,
         zone_id: usize,
-        colors: Vec<Color>,
+        colors: impl IntoIterator<Item = Color>,
     ) -> OpenRgbResult<()> {
-        let cmd = self.get_controller_mut(controller_id)?;
-        cmd.add_set_zone_leds(zone_id, colors)
+        let cmd = self.get_cmd_mut(controller_id)?;
+        cmd.set_zone_leds(zone_id, colors)
     }
 
     /// Add a command to update a single LED in a zone in a controller.
@@ -138,15 +131,15 @@ impl<'a> UpdateLedCommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn add_update_zone_led(
+    pub fn set_controller_zone_led(
         &mut self,
         controller_id: impl ControllerIndex,
         zone_id: usize,
         led_idx: usize,
         color: Color,
     ) -> OpenRgbResult<()> {
-        let cmd = self.get_controller_mut(controller_id)?;
-        cmd.add_set_zone_led(zone_id, led_idx, color)
+        let cmd = self.get_cmd_mut(controller_id)?;
+        cmd.set_zone_led(zone_id, led_idx, color)
     }
 
     /// Add a command to update a segment in a zone in a controller.
@@ -154,31 +147,32 @@ impl<'a> UpdateLedCommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn add_update_segment(
+    pub fn set_controller_segment_leds(
         &mut self,
         controller_id: impl ControllerIndex,
         zone_id: usize,
         segment_id: usize,
-        colors: Vec<Color>,
+        colors: impl IntoIterator<Item = Color>,
     ) -> OpenRgbResult<()> {
-        let cmd = self.get_controller_mut(controller_id)?;
-        cmd.add_set_segment_leds(zone_id, segment_id, colors)
+        let cmd = self.get_cmd_mut(controller_id)?;
+        cmd.set_segment_leds(zone_id, segment_id, colors)
     }
 }
 
 /// A command to update the LEDs in a controller.
+/// Commands must be executed using `[Self::execute()]`. Use `[Controller::cmd()]` to create a command.
 ///
 /// When executed, all commands are combined into a single `UpdateCommand::Controller`,
 /// meaning only a single update is actually sent to the controller.
 ///
 /// When two commands write to the same LED, the last command will overwrite the previous one.
 #[derive(Debug)]
-pub struct UpdateLedCommand<'a> {
+pub struct Command<'a> {
     controller: &'a Controller,
     colors: Vec<Color>,
 }
 
-impl<'a> UpdateLedCommand<'a> {
+impl<'a> Command<'a> {
     pub(crate) fn new(controller: &'a Controller) -> Self {
         Self {
             controller,
@@ -194,47 +188,44 @@ impl<'a> UpdateLedCommand<'a> {
     }
 
     /// Adds a command to update a single LED in this controller.
-    pub fn add_set_led(&mut self, led_id: usize, color: Color) -> OpenRgbResult<()> {
-        self.add_command(UpdateCommand::Single {
-            controller_id: self.controller.id(),
-            led_id,
-            color,
-        })
+    pub fn set_led(&mut self, led_id: usize, color: Color) -> OpenRgbResult<()> {
+        self.add_command(SetLedCommand::Single { led_id, color })
     }
 
     /// Adds a command to update multiple LEDs in this controller.
-    pub fn add_set_leds(&mut self, colors: Vec<Color>) -> OpenRgbResult<()> {
-        self.add_command(UpdateCommand::Controller {
-            controller_id: self.controller.id(),
-            colors,
+    pub fn set_leds(&mut self, colors: impl IntoIterator<Item = Color>) -> OpenRgbResult<()> {
+        self.add_command(SetLedCommand::Controller {
+            colors: colors.into_iter().collect(),
         })
     }
 
     /// Adds a command to update a single LED in a zone in this controller.
-    pub fn add_set_zone_led(
+    pub fn set_zone_led(
         &mut self,
         zone_id: usize,
         led_idx: usize,
         color: Color,
     ) -> OpenRgbResult<()> {
-        self.add_command(UpdateCommand::Single {
-            controller_id: self.controller.id(),
+        self.add_command(SetLedCommand::Single {
             led_id: self.controller.get_zone_led_offset(zone_id)? + led_idx,
             color,
         })
     }
 
     /// Adds a command to update multiple LEDs in a zone in this controller.
-    pub fn add_set_zone_leds(&mut self, zone_id: usize, colors: Vec<Color>) -> OpenRgbResult<()> {
-        self.add_command(UpdateCommand::Zone {
-            controller_id: self.controller.id(),
+    pub fn set_zone_leds(
+        &mut self,
+        zone_id: usize,
+        colors: impl IntoIterator<Item = Color>,
+    ) -> OpenRgbResult<()> {
+        self.add_command(SetLedCommand::Zone {
             zone_id,
-            colors,
+            colors: colors.into_iter().collect(),
         })
     }
 
     /// Adds a command to update a single led in a segment in a zone in this controller.
-    pub fn add_set_segment_led(
+    pub fn set_segment_led(
         &mut self,
         zone_id: usize,
         segment_id: usize,
@@ -247,46 +238,27 @@ impl<'a> UpdateLedCommand<'a> {
                 .get_zone(zone_id)?
                 .get_segment(segment_id)?
                 .offset();
-        self.add_command(UpdateCommand::Single {
-            controller_id: self.controller.id(),
-            led_id,
-            color,
-        })
+        self.add_command(SetLedCommand::Single { led_id, color })
     }
 
     /// Adds a command to update multiple LEDs in a segment in a zone in this controller.
-    pub fn add_set_segment_leds(
+    pub fn set_segment_leds(
         &mut self,
         zone_id: usize,
         segment_id: usize,
-        colors: Vec<Color>,
+        colors: impl IntoIterator<Item = Color>,
     ) -> OpenRgbResult<()> {
-        self.add_command(UpdateCommand::Segment {
-            controller_id: self.controller.id(),
+        self.add_command(SetLedCommand::Segment {
             zone_id,
             segment_id,
-            colors,
+            colors: colors.into_iter().collect(),
         })
     }
 
-    /// Extend this command with multiple other `UpdateCommand`s.
-    pub fn extend_with(
-        &mut self,
-        commands: impl IntoIterator<Item = UpdateCommand>,
-    ) -> OpenRgbResult<&mut Self> {
-        for cmd in commands {
-            self.add_command(cmd)?;
-        }
-        Ok(self)
-    }
-
     /// Adds an `UpdateCommand` to this command.
-    pub fn add_command(&mut self, cmd: UpdateCommand) -> OpenRgbResult<()> {
+    fn add_command(&mut self, cmd: SetLedCommand) -> OpenRgbResult<()> {
         match cmd {
-            UpdateCommand::Controller {
-                controller_id: _,
-                colors,
-            } => {
+            SetLedCommand::Controller { colors } => {
                 if colors.len() > self.controller.num_leds() {
                     tracing::warn!(
                         "Controller {} was given {} colors, while its length is {}. This might become a hard error in the future.",
@@ -298,11 +270,7 @@ impl<'a> UpdateLedCommand<'a> {
 
                 self.set_colors(0, &colors)?;
             }
-            UpdateCommand::Zone {
-                controller_id: _,
-                zone_id,
-                colors,
-            } => {
+            SetLedCommand::Zone { zone_id, colors } => {
                 let zone = self.controller.get_zone(zone_id)?;
                 if colors.len() >= zone.num_leds() {
                     tracing::warn!(
@@ -318,8 +286,7 @@ impl<'a> UpdateLedCommand<'a> {
                 let len = colors.len().min(zone.num_leds());
                 self.set_colors(offset, &colors[..len])?;
             }
-            UpdateCommand::Segment {
-                controller_id: _,
+            SetLedCommand::Segment {
                 zone_id,
                 segment_id,
                 colors,
@@ -340,11 +307,7 @@ impl<'a> UpdateLedCommand<'a> {
                 let offset = zone.offset() + seg.offset();
                 self.set_colors(offset, &colors)?;
             }
-            UpdateCommand::Single {
-                controller_id: _,
-                led_id,
-                color,
-            } => {
+            SetLedCommand::Single { led_id, color } => {
                 if led_id >= self.controller.num_leds() {
                     tracing::warn!(
                         "LED id {} is out of bounds for controller {} with {} LEDs",